requirements.txt
line-bot-sdk
fastapi
uvicorn[standard]
gunicorn
google-cloud-secret-manager
google-cloud-firestore
uvicorn-worker
